+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "IP mask calculating".
 - Tests for "Domino puzzle".
 - Tests for "Goldbach’s conjecture".
+- Tests for "Let's go to the cinema!".
 
 ## [1.16.0] - 2022-12-31
 ### Added